move editor inside toolbox
