+++ index.js
@@ -11,15 +11,14 @@ import { getCatalogLayerById }                 from './utils/getCatalogLayerById
 import { OpenFormStep }                        from './actions/open-form';
 import { AddFeatureStep }                      from './actions/add-feature';
 import { ToolBox }                             from './g3w-toolbox';
-import { Collection }                          from './g3w-collection';
 
 const { G3W_FID }                              = g3wsdk.constant;
-const { ApplicationState, G3WObject }          = g3wsdk.core;
+const { ApplicationState }                     = g3wsdk.core;
 const _                                        = g3wsdk.core.i18n.t;
 const { Layer, LayersStore }                   = g3wsdk.core.layer;
 const { Feature }                              = g3wsdk.core.layer.features;
 const { Plugin, PluginService }                = g3wsdk.core.plugin;
-const { XHR, noop, cloneDeep }                 = g3wsdk.core.utils;
+const { XHR, noop }                            = g3wsdk.core.utils;
 const { GUI }                                  = g3wsdk.gui;
 const { Panel }                                = g3wsdk.gui.vue;
 const { Server: serverErrorParser }            = g3wsdk.core.errors.parsers;
@@ -28,8 +27,6 @@ const {
   getScaleFromResolution,
   getResolutionFromScale,
 }                                              = g3wsdk.ol.utils;
-
-const is_defined = d => undefined !== d;
 
 Object
   .entries({
@@ -63,12 +60,13 @@ new (class extends Plugin {
      */
     this.state = {
       open:                false, // check if panel is open or not
-      editors:             {},
-      sessions:            {},    // store all sessions
       toolboxes:           [],
       toolboxselected:     null,
-      /** @since g3w-client-plugin-editing@v3.6.2 */
       showselectlayers:    true,  // whether to show selected layers on editing panel
+      layers: {},                // editable layers (vector)
+      features: {},              // edited features (local)
+      lock_ids: {},              // locked features
+      loaded_ids: {},            // Ids of features loaded by current user
       message:             null,
       relations:           [],
       layers_in_error:     false,
@@ -89,10 +87,6 @@ new (class extends Plugin {
           error: () => {}        // function executed after commit changes error
         }
       },
-      layers: {},                // editable layers (vector)
-      features: {},              // edited features (local)
-      lock_ids: {},              // locked features
-      loaded_ids: {},            // Ids of features loaded by current user
       events:              {
         'start-editing':         {},
         'show-relation-editing': {},
@@ -196,12 +190,7 @@ new (class extends Plugin {
 
     // add editing layer store to mapstoreregistry
     ApplicationState.layers['editing'] = new LayersStore({ id: 'editing', queryable: false, catalog: false });
-
-    this.state.layers     = {};
-    this.state.toolboxes  = [];
     
-    let count = 0;
-
     // loop over editable layers
     (await Promise.allSettled(
       getCatalogLayers({ EDITABLE: true }, { TOC_ORDER : true })
@@ -209,17 +198,16 @@ new (class extends Plugin {
         .map(async (layer, index) => {
           try {
 
-          if (!layer.isEditable()) {
-            return null;
-          }
+            if (!layer.isEditable()) {
+              return null;
+            }
           
-          // get layer editing config (from server)
-          try {
+            // get layer editing config (from server)
             const {
-            vector,
-            constraints = {},
-            capabilities,
-          } = await layer.getProvider('data').getConfig();
+              vector,
+              constraints = {},
+              capabilities,
+            } = await layer.getProvider('data').getConfig();
 
             layer.state.editing =  {
               started:  false,
@@ -233,11 +221,11 @@ new (class extends Plugin {
               format:                      vector.format,
               constraints,
               capabilities:                capabilities || ['add_feature', 'change_feature', 'change_attr_feature', 'delete_feature' ], // default editing capabilities
-              form:                        { perc: null },                                                      // set editing form `perc` to null at beginning
-              style:                       vector.style,                                                        // get vector layer style
-              geometrytype:                vector.geometrytype,                                                 // whether is a vector layer,
-              visible:                     (vector.editing || { visible: true }).visible,                       //@since 3.11.0 let know if layer should be editable directly (true) or through relation layer (false)
-              layer_style:                 (vector.editing || { layer_style: null }).layer_style,               // @since v4.0.0 check if has a layer style to for editing form
+              form:                        { perc: null },                                        // set editing form `perc` to null at beginning
+              style:                       vector.style,                                          // get vector layer style
+              geometrytype:                vector.geometrytype,                                   // whether is a vector layer,
+              visible:                     (vector.editing || { visible: true }).visible,         // @since 3.11.0 let know if layer should be editable directly (true) or through relation layer (false)
+              layer_style:                 (vector.editing || { layer_style: null }).layer_style, // @since v4.0.0 check if has a layer style to for editing form
             };
 
             // set vector layer color 
@@ -246,202 +234,9 @@ new (class extends Plugin {
             }
 
             layer.state.editing.ready = true;
-          } catch(e) {
-            console.warn(e);
-          }
 
-          // set editing layer
-          let editing_layer = Layer.LayerTypes.IMAGE === layer.getType()
-            ? new g3wsdk.core.layer.VectorLayer(layer.config)
-            : layer;
-
-          const suffixUrl = `${ApplicationState.project.getType()}/${ApplicationState.project.getId()}/${layer.getId()}/`;
-          const vectorUrl =  ApplicationState.project.state.vectorurl;
-
-          this.state.features[layer.getId()]   = new Collection(Layer.LayerTypes.TABLE !== layer.getType());
-          this.state.lock_ids[layer.getId()]   = [];
-          this.state.loaded_ids[layer.getId()] = [];
-
-          /**
-           * ORIGINAL SOURCE: g3w-client-plugin-editing/g3wsdk/editing/editor.j@v4.0.0
-           * ORIGINAL SOURCE: g3w-client/src/map/layers/featuresstore.js@v4.0.0
-           * ORIGINAL SOURCE: g3w-client/src/app/core/layers/features/olfeaturesstore.js@v3.10.2
-           */
-          const editor = editing_layer._editor = Object.assign(new G3WObject, {
-
-            /** Filter to getFeaturerequest */
-            _filter: { bbox: null },
-            /** @type { Boolean } true, mean all features of layer are get (e.g. Table layer) */
-            _allfeatures: false,
-            /** Referred layer */
-            _layer:     layer,
-            /** Original features (from server) */
-            _features: [],
-            /** @type { boolean } Whether editor is active or not */
-            _started: false,
-            urls: {
-              editing:     `${vectorUrl}editing/${suffixUrl}`,
-              commit:      `${vectorUrl}commit/${suffixUrl}`,
-              config:      `${vectorUrl}config/${suffixUrl}`,
-              unlock:      `${vectorUrl}unlock/${suffixUrl}`,
-            },
-            /** Store editing features */
-            _featuresstore: Object.assign(new G3WObject, {
-                setters: {
-                  addFeatures: (feats = []) => feats.forEach(f => editor._featuresstore.addFeature(f)),
-                  removeFeature: f => this.state.features[layer.getId()].remove(f),
-                  updateFeature: f => this.state.features[layer.getId()].update(f),
-                },
-                clear:                     () => this.state.features[layer.getId()].clear(),
-                addFeature:                f => this.state.features[layer.getId()].add(f),
-                clone:                     () => cloneDeep(editor._featuresstore),
-                getFeatureById:            () => this.state.features[layer.getId()].getArray().find(f => id == f.getId()),
-                readFeatures:              () => this.state.features[layer.getId()].getArray(),
-                getLength:                 () => this.state.features[layer.getId()].getArray().length,
-                getFeaturesCollection:     () => this.state.features[layer.getId()]._store,
-                setFeatures:               (feats = []) => { this.state.features[layer.getId()].clear(); editor._featuresstore.addFeatures(feats); },
-            }),
-            setters: {
-              save:                       () => layer.save(),
-              addFeature:                 f => editor._featuresstore.addFeature(f),
-              updateFeature:              f => editor._featuresstore.updateFeature(f),
-              deleteFeature:              f => editor._featuresstore.deleteFeature(f),
-              setFeatures:               (f = []) => editor._featuresstore.setFeatures(f),
-              getFeatures:               (l, o, p) => this.__getFeatures(layer.getId(), o, p),
-              featuresLockedByOtherUser: f => {},
-            },
-            addFeature:          f => editor._featuresstore.addFeature(f),
-            isStarted:           () => editor._started,
-            getLockIds:          () => this.state.lock_ids[layer.getId()],
-            getEditingSource:    () => editor._featuresstore,
-            getSource:           () => layer.getSource(),
-            getLayer:            () => layer,
-            rollback:            (c = []) => this.__setChanges(layer.getId(), c, true),
-            readFeatures:        () => editor._features,
-            readEditingFeatures: () => editor._featuresstore.readFeatures(),
-            commit:              c => this.__commitToEditor(layer.getId(), c),
-            start:               o => this.__startEditor(layer.getId(), o),
-            stop:                () => this.__stopEditor(layer.getId()),
-            clear:               () => this.__clearEditor(layer.getId()),
-          });
-
-          // clone editable layer
-          if (Layer.LayerTypes.TABLE === editing_layer.getType()) {
-            editing_layer = editing_layer.clone(); 
-          }
-
-          this.state.editors[layer.getId()]            = editing_layer.getEditor(); 
-          this.state.layers[layer.getId()]             = editing_layer;
-          this.state.uniqueFieldsValues[layer.getId()] = {};
-
-          /**
-           * attach layer widgets event: get data from api when a field of a layer
-           * is related to a wgis form widget (ex. relation reference, value map, etc..)
-           */
-          editing_layer
-            .getEditingFields()
-            .filter(field => field.input && 'select_autocomplete' === field.input.type && !field.input.options.filter_expression && !field.input.options.usecompleter)
-            /** @TODO need to avoid to call the same fnc to same event many times to avoid waste server request time */
-            .forEach(field => {
-              this.state.events['start-editing'][layer.getId()] = this.state.events['start-editing'][layer.getId()] || [];
-              this.state.events['start-editing'][layer.getId()].push(async () => {
-                // remove all values
-                field.input.options.loading.state = 'loading';
-                field.input.options.values        = [];
-
-                const relationLayer = field.input.options.layer_id && getCatalogLayerById(field.input.options.layer_id);
-                const has_filter    = ([undefined, null].includes(field.input.options.filter_fields || []) || 0 === (field.input.options.filter_fields || []).length);
-
-                try {
-
-                  // relation reference widget + no filter set
-                  if (field.input.options.relation_reference && has_filter) {
-                    const response = await editing_layer.getFilterData({ fformatter: field.name }); // get data with fformatter
-                    if (response && response.data) {
-                      // response data is an array ok key value objects
-                      field.input.options.values.push(...response.data.map(([value, key]) => ({ key, value })));
-                      field.input.options.loading.state = 'ready';
-                      this.fireEvent('autocomplete', { field, data: [response.data] });
-                      return field.input.options.values;
-                    }
-                  }
-
-                  // value map widget
-                  if (relationLayer) {
-                    //ordering by value or key depend on orderbyvalue Boolean value
-                    const response = await relationLayer.getDataTable({ ordering: field.input.options.orderbyvalue ? field.input.options.value : field.input.options.key });
-                    if (response && response.features) {
-                      field.input.options.values.push(...(response.features || []).map(feature => ({
-                        key:   feature.properties[field.input.options.value],
-                        value: feature.properties[field.input.options.key],
-                      })));
-                      field.input.options.loading.state = 'ready';
-                      this.fireEvent('autocomplete', { field, features: response.features })
-                      return field.input.options.values;
-                    }
-                  }
-
-                  /** @TODO check if deprecated */
-                  const features        = [];
-                  field.input.options.loading.state = 'ready';
-                  this.fireEvent('autocomplete', { field, features });
-                  return features;
-
-                } catch (e) {
-                  console.warn(e);
-                  field.input.options.loading.state = 'error';
-                  return Promise.reject(e);
-                }
-              });
-            });
-
-            this.state.sessions[layer.getId()] = null;
-
-            /**
-             * set 1:1 relation fields editable
-             * 
-             * Check if layer has relation 1:1 (type ONE) and if fields
-             *
-             * belongs to relation where child layer is editable
-             *
-             * @since g3w-client-plugin-editing@v3.7.0
-             */
-            getCatalogLayerById(layer.getId())
-              .getRelations()
-              .getArray()
-              .filter(relation => 'ONE' === relation.getType() && layer.getId() === relation.getFather()) // 'ONE' == join 1:1 + father layerId is a father of relation
-              .forEach(relation => {
-                const isChildEditable = undefined !== this.getLayerById(relation.getChild());        // check if child layerId is editable (in editing)
-                this
-                  .getLayerById(relation.getFather())
-                  .getEditingFields()
-                  .filter(f => f.vectorjoin_id && f.vectorjoin_id === relation.getId())  // father layer fields (in editing)
-                  .forEach(f => { f.editable = (f.editable && isChildEditable); });      // current editable boolean value + child editable layer
-              });
-
-            // Set editing layer color and toolbox style
-            if (!editing_layer.getColor()) {
-              editing_layer.setColor(editing_layer.isGeoLayer() ? [
-                "#C43C39", "#d95f02", "#91522D", "#7F9801", "#0B2637",
-                "#8D5A99", "#85B66F", "#8D2307", "#2B83BA", "#7D8B8F",
-                "#E8718D", "#1E434C", "#9B4F07", '#1b9e77', "#FF9E17",
-                "#7570b3", "#204B24", "#9795A3", "#C94F44", "#7B9F35",
-                "#373276", "#882D61", "#AA9039", "#F38F3A", "#712333",
-                "#3B3A73", "#9E5165", "#A51E22", "#261326", "#e4572e",
-                "#29335c", "#f3a712", "#669bbc", "#eb6841", "#4f372d",
-                "#cc2a36", "#00a0b0", "#00b159", "#f37735", "#ffc425",
-              ][count++ % 40] : '#fff');
-            }
-
-            //create toolbox
-            const toolbox = new ToolBox(editing_layer, [
-              ...editing_layer.getChildren(),
-              ...editing_layer.getFathers()].filter(id => this.getLayerById(id))
-            );
-            //Use index to mantain TOC layer order
-            this.state.toolboxes[index] = toolbox;
-            this.state.sessions[layer.getId()] = toolbox.getSession(); // add session
-
+            // NB: sorted by "index" to keep TOC order
+            this.state.toolboxes[index] = new ToolBox(layer, this);
           } catch (e) {
             this.state.layers_in_error = true;
             console.warn(e);
@@ -634,7 +429,6 @@ new (class extends Plugin {
    */
   addToolBox(toolbox) {
     this.state.toolboxes.push(toolbox);
-    this.state.sessions[toolbox.getId()] = toolbox.getSession(); // add session
   }
 
   /**
@@ -741,7 +535,7 @@ new (class extends Plugin {
    * @since g3w-client-plugin-editing@v3.7.0
    */
   getSessionById(id) {
-    return this.state.sessions[id];
+    return this.getToolBoxById(id).getSession();
   }
 
   /**
@@ -1453,423 +1247,6 @@ new (class extends Plugin {
    */
   getActiveTool() {
     return this.getToolBoxes().filter(t => t.getActiveTool())[0];
-  }
-
-  /**
-   * @since g3w-client-plugin-editing@v4.1.0
-   */
-  editor(layerId) {
-    return this.state.editors[layerId];
-  }
-
-  /**
-   * ORIGINAL SOURCE: g3w-client-plugin-editing/g3wsdk/editing/editor.j@v4.0.0
-   * 
-   * Get features from server method.
-   * Used when vector Layer's bbox is contained into an already requested bbox (so no a new request is done).
-   *
-   * @param { number[] } options.filter.bbox bounding box Array [xmin, ymin, xmax, ymax]
-   *
-   * @returns { boolean } whether can perform a server request
-   * 
-   * @since g3w-client-plugin-editing@v4.1.0
-   */
-  async __getFeatures(layerId, options = {}, params = {}) {
-    const editor = this.editor(layerId);
-
-    // skip is not onlien or all features of layers are already got
-    if (!ApplicationState.online || editor._allfeatures) {
-      return Promise.resolve();
-    }
-
-    let doRequest = true; // default --> perform request
-
-    const { bbox } = options.filter || {};
-    //check if bbox options filter (bbox of a current map) is passed and is a vector layer
-    const is_vector = bbox && Layer.LayerTypes.VECTOR === editor.getLayer().getType();
-
-    // first request --> need to perform request
-    if (is_vector && null === editor._filter.bbox) {
-      editor._filter.bbox = bbox;                                                      // store bbox
-      doRequest         = true;
-    }
-
-    // subsequent requests --> check if bbox is contained into an already requested bbox
-    else if (is_vector) {
-      //Boolean - Check if features are already got inside bbox
-      const is_cached = ol.extent.containsExtent(editor._filter.bbox, bbox);
-      if (!is_cached) {
-        editor._filter.bbox = ol.extent.extend(editor._filter.bbox, bbox);
-      }
-      doRequest = !is_cached;
-    }
-
-    if (!doRequest) {
-      return;
-    }
-
-    const url = editor.urls.editing;
-    try {
-      let response;
-      if (!options.filter) {
-        response = await XHR.post({
-          url,
-          data:        JSON.stringify(params),
-          contentType: 'application/json',
-        });
-      } else if (is_defined(options.filter.bbox)) { // bbox filter
-        response = await XHR.post({
-          url,
-          data: JSON.stringify({
-            ...params,
-            in_bbox:     options.filter.bbox.join(','),
-            filtertoken: editor.getLayer().getFilterToken(),
-          }),
-          contentType: 'application/json',
-        })
-      } else if (is_defined(options.filter.fid)) { // fid filter
-        const { fid, relation } = options.filter.fid;
-        response = await XHR.post({
-          url: `${editor.urls.editing}?relationonetomany=${relation.id}|${fid}`,
-          contentType: 'application/json',
-          data:        JSON.stringify({ formatter: 1 }),
-        });
-      } else if (options.filter.field) {
-        response = await XHR.post({
-          url,
-          data:        JSON.stringify({ 
-            ...params,
-            ...options.filter,
-          }),
-          contentType: 'application/json',
-        })
-      } else if (is_defined(options.filter.fids)) {
-        response = await XHR.post({
-          url,
-          data:   JSON.stringify({
-            ...params,
-            ...options.filter,
-          }),
-          contentType: 'application/json',
-        })
-      } else if (is_defined(options.filter.nofeatures)) {
-        response = await XHR.post({
-          url,
-          data: JSON.stringify({
-            ...params,
-            field: `${options.filter.nofeatures_field || 'id'}|eq|__G3W__NO_FEATURES__`
-          }),
-          contentType: 'application/json',
-        })
-      }
-
-      // invalid response
-      if (!response.result) {
-        return;
-      }
-
-      const { data, count }       = response.vector;
-      const { featurelocks = [] } = response;
-      const lockIds               = featurelocks.map(lk => lk.featureid);
-      const dataProjection = 'NoGeometry' === response.vector.geometrytype ? null : editor.getLayer().getCrs();
-      let features   = [];
-
-      try {
-
-        features = (new ol.format.GeoJSON({
-          geometryName:      'geometry',
-          dataProjection,
-          featureProjection: dataProjection,
-        }))
-        .readFeatures('string' === typeof data ? JSON.parse(data) : data)
-        .filter(f => lockIds.includes(`${f.getId()}`))
-        .map(feature => new Feature({ feature }));
-
-        //if no features locks mean another user locks all feature requests
-        if (0 === featurelocks.length || count > features.length) {
-          //It means that another user locks these features
-          editor.featuresLockedByOtherUser(features);
-        }
-        //get already loaded feature id locked by current user
-        const fids = lockIds.map(({ featureid }) => featureid);
-        featurelocks
-          .filter(({ featureid }) => !fids.includes(featureid)) //exclude features already locked by current user
-          .forEach(fl => this.state.lock_ids[layerId].push(fl)) //update lockIds based on a featurelocks array from response
-
-        //store features locked by another user
-        const lockFeatures = [];
-
-        //Store features to add to layers source
-        features = features.filter(f => {
-          //get feature id
-          const featureId = f.getId();
-          //check if feature id is locked features
-          //it means that is not locked by another user.
-          if (featurelocks.find(({ featureid }) => featureId == featureid)) {
-            //check if feature is not yet added for the current user
-            if (!this.state.loaded_ids[layerId].includes(featureId)) {
-              this.state.loaded_ids[layerId].push(featureId);
-              return true;
-            } else {
-              return false; //feature locked by the current user
-            }
-          } else {
-            lockFeatures.push(f);
-            return false; //feature locked by another user
-          }
-        });
-
-    } catch (e) {
-      console.warn(e);
-    }
-
-    editor.readFeatures().push(...features); // add features to original features 
-    
-    // add features from server to editing features store (cloned from original)
-    editor.getEditingSource().addFeatures((features || []).map(f => f.clone()));
-
-    //set all features to true if no filter is set (e.g., Table layer)
-    editor._allfeatures = !options.filter;
-
-    return features;
-    } catch(e) {
-      console.warn(e);
-      return Promise.reject({ message: _("info.server_error")});
-    }
-
-  }
-
-  /**
-   * ORIGINAL SOURCE: g3w-client/src/services/editing.js@v3.9.1
-   * 
-   * Apply changes to source features (undo/redo)
-   * 
-   * @param items
-   * @param { boolean } reverse whether change to opposite
-   * 
-   * @since g3w-client-plugin-editing@v4.1.0
-   */
-  __setChanges(layerId, items = [], reverse = true) {
-    const editor = this.editor(layerId);
-    /** known actions */
-    const Actions = {
-      'add':    { fnc: 'addFeature',    opposite: 'delete' },
-      'delete': { fnc: 'removeFeature', opposite: 'add'    },
-      'update': { fnc: 'updateFeature', opposite: 'update' },
-    };
-    items.forEach(item => {
-      if (reverse) {
-        item.feature[Actions[item.feature.getState()].opposite]();
-      }
-      // get method from object
-      //@since 3.9.1 need to clone it otherwise it replace
-      editor.getEditingSource()[Actions[item.feature.getState()].fnc](item.feature.clone());
-    });
-  }
-
-  /**
-   * ORIGINAL SOURCE: g3w-client-plugin-editing/g3wsdk/editing/editor.j@v4.0.0
-   * 
-   * Run after server has applied changes to origin resource
-   *
-   * @param commit commit items
-   *
-   * @returns jQuery promise
-   * 
-   * @since g3w-client-plugin-editing@v4.1.0
-   */
-  async __commitToEditor(layerId, commit) {
-
-    const editor = this.editor(layerId);
-    
-    let relations = [];
-
-    // check if there are commit relations binded to new feature
-    if (commit.add.length) {
-      relations = Object
-        .keys(commit.relations)
-        .map(relationId => {
-          const relation = editor.getLayer().getRelations().getRelationByFatherChildren(layerId, relationId);
-          return {
-            [relationId]: {
-              ids: [                                                  // ids of "added" or "updated" relations
-                ...commit.relations[relationId].add.map(r => r.id),   // added
-                ...commit.relations[relationId].update.map(r => r.id) // updated
-              ],
-              fatherField: relation.getFatherField(), // father Fields <Array>
-              childField:  relation.getChildField()    // child Fields <Array>
-            }
-          };
-        });
-    }
-
-    // commit items
-    let response;
-
-    try {
-      commit.lockids = this.state.lock_ids[layerId];
-      response = await XHR.post({
-        url:         editor.urls.commit,
-        data:        JSON.stringify(commit),
-        contentType: 'application/json',
-      });
-    } catch(e) {
-      console.warn(e);
-      response = Promise.reject();
-    }
-
-    // sync selection filter features
-    if (response?.result) {
-      try {
-        const layer = getCatalogLayerById(layerId);
-        //if layer has geometry
-        if (layer.isGeoLayer()) {
-          commit.update.forEach(({ id, geometry } = {}) => {
-            if (layer.getOlSelectionFeature(id)) {
-              const selected = layer.getOlSelectionFeature(id);
-              if (selected) {
-                selected.feature = geometry;
-                GUI.getService('map').setSelectionFeatures('update', { feature: geometry });
-              }
-            }
-          });
-        }
-        commit.delete.forEach(id => {
-          if (layer.hasSelectionFid(id)) {
-            layer.excludeSelectionFid(id);
-          }
-        })
-      } catch(e) {
-        console.warn(e);
-      }
-    }
-
-    // skip when no response and response.result is false
-    if (!(response && response.result)) {
-      return response;
-    }
-
-    //Loop on new features saved on server
-    // clientid - temporary id of new feature
-    // id - id saved on server (autogenerate, next value) to subtituite to clientid feature id
-    // properties - properties of feature returned by server
-    response.response.new.forEach(({ clientid, id, properties } = {}) => {
-      //get feature from current layer in editing
-      const feature  = editor.getEditingSource().getFeatureById(clientid);
-      // set new id
-      feature.setId(id);
-      //set properties
-      feature.setProperties(properties);
-      //Loop on eventual relation updated or created
-      relations.forEach(r => {         // handle relations (if provided)
-        Object
-          .entries(r)
-          .forEach(([ id, opts = {}]) => { // id - relation layer id, opts - Object contain relation properties
-            //get the editing source of relation layer
-            const source = ToolBox.get(id).getSession().getEditor().getEditingSource();
-            // handle value to relation field saved on server
-            (opts.ids || []).forEach(id => {
-              const rFeature = source.getFeatureById(id);
-              if (rFeature) {
-                opts.fatherField.forEach((ff, i) => {// loop relation ids
-                  rFeature.set(opts.childField[i], feature.get(ff))  // set father feature `value` and `name`
-                })
-              }
-            })
-          });
-      });
-
-    });
-
-    //@since 3.9.0 take in account update properties returned by server (Useful in case of media input changes)
-    (response.response.update || []).forEach(({ id, properties } = {}) => {
-      //get feature from current layer in editing
-      const feature  = editor.getEditingSource().getFeatureById(id);
-      //set properties
-      feature.setProperties(properties);
-      //Loop on eventual relation updated or created
-      relations.forEach(r => {         // handle relations (if provided)
-        Object
-          .entries(r)
-          .forEach(([ id, opts = {}]) => { // id - relation layer id, opts - Object contain relation properties
-            //get the editing source of relation layer
-            const source = ToolBox.get(id).getSession().getEditor().getEditingSource();
-            // handle value to relation field saved on server
-            (opts.ids || []).forEach(id => {
-              const rFeature = source.getFeatureById(id);
-              if (rFeature) {
-                opts.fatherField.forEach((ff, i) => {// loop relation ids
-                  rFeature.set(opts.childField[i], feature.get(ff))  // set father feature `value` and `name`
-                })
-              }
-            })
-          });
-      });
-
-    });
-
-    const features = editor.readEditingFeatures();
-
-    features.forEach(f => f.clearState());          // reset state of the editing features (update, new etc..)
-
-    editor.getLayer().setFeatures([...features]);         // substitute layer features with actual editing features ("cloned" to prevent layer actions duplicates, eg. addFeatures)
-
-    // add lock ids
-    this.state.lock_ids[layerId] = [...new Set(this.state.lock_ids[layerId].concat(...response.response.new_lockids))]
-    this.state.lock_ids[layerId].forEach(({ featureid }) => this.state.loaded_ids[layerId].push(featureid));
-
-    return response;
-  }
-
-  /**
-   * ORIGINAL SOURCE: g3w-client-plugin-editing/g3wsdk/editing/editor.j@v4.0.0
-   * 
-   * start editing
-   * 
-   * @since g3w-client-plugin-editing@v4.1.0
-   */
-  async __startEditor(layerId, options = {}) {
-    const editor   = this.editor(layerId)
-    const features = await editor.getFeatures(options);     // load layer features based on filter type
-    editor._started  = true;                                 // if all ok set to started
-    return features;                                      // features are already inside featuresstore
-  }
-
-  /**
-   * ORIGINAL SOURCE: g3w-client-plugin-editing/g3wsdk/editing/editor.j@v4.0.0
-   * 
-   * stop editor (unlock)
-   * 
-   * @since g3w-client-plugin-editing@v4.1.0
-   */
-  async __stopEditor(layerId) {
-    const editor     = this.editor(layerId)
-    const { result } = await XHR.post({ url: editor.urls.unlock });
-    editor.clear();
-    return result;
-  }
-
-  /**
-   * ORIGINAL SOURCE: g3w-client-plugin-editing/g3wsdk/editing/editor.j@v4.0.0
-   * 
-   * @since g3w-client-plugin-editing@v4.1.0 
-   */
-  __clearEditor(layerId) {
-    const editor = this.editor(layerId)
-
-    editor._started     = false;
-    editor._filter.bbox = null;
-    editor._allfeatures = false;
-
-    editor._features               = []; // clear features collection
-    this.state.lock_ids[layerId]   = [];
-    this.state.loaded_ids[layerId] = [];
-    editor.getEditingSource().clear();
-
-    // vector layer
-    if (Layer.LayerTypes.VECTOR === editor.getLayer().getType()) {
-      editor.getLayer().resetEditingSource(editor.getEditingSource().getFeaturesCollection());
-    }
   }
 
   /**
